+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/component/ImplicitFile.kt
@@ -1,5 +1,6 @@
 package eu.vendeli.tgbot.types.component
 
+import eu.vendeli.tgbot.utils.common.safeCast
 import eu.vendeli.tgbot.utils.serde.ToStringSerializer
 import kotlinx.serialization.Serializable
 
@@ -31,5 +32,5 @@ sealed class ImplicitFile {
         ) = FileId(file)
     }
 
-    internal object Serde : ToStringSerializer<ImplicitFile>({ (this as? FileId)?.file.toString() })
+    internal object Serde : ToStringSerializer<ImplicitFile>({ this.safeCast<FileId>()?.file.toString() })
 }